requirements.txt
Django==1.5.1
South==0.8.1
amqp==1.0.12
anyjson==0.3.3
billiard==2.7.3.31
bunch==1.0.1
celery==3.0.21
celery-with-redis==3.0
distribute==0.7.3
django-annoying==0.7.7
django-celery==3.0.17
django-celery-with-redis==3.0
django-kombu==0.9.4
django-oauth-tokens==0.2.6
kombu==2.5.12
pyfacegraph
python-dateutil==2.1
pytz==2013b
redis==2.7.6
requests==1.2.3
six==1.3.0
tyoi.OAuth2==0.2.1
ujson==1.33
urllib3==1.6
wsgiref==0.1.2
psycopg2